# swagger/static/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Safe Foods - OpenAPI 3.0",
    "description": "This is a basic Flask API that serves up and edits a database which contains data for a group of restaurants, each with associated dishes. Those two tables alongside a list of users covers the entire app. This is built as a RESTful API with all endpoints residing in `main.py`.\nYou can find out more about this project on its associated GitHub repo, as well as its DB Schema and UI repos. This is not associated with any real app, paid or otherwise, and is purely for personal and demonstration purposes.",
    "contact": {
      "email": "connorjamesl@hotmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "1.0.0"
  },
  "paths": {
    "/restaurants": {
      "get": {
        "tags": [
          "restaurant"
        ],
        "summary": "Read data on all restaurants",
        "description": "Runs a `SELECT *` query against all data in the restaurants table.",
        "operationId": "getRestaurants",
        "responses": {
          "200": {
            "description": "Successful Read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Restaurants"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "restaurant"
        ],
        "summary": "Add a new restaurant.",
        "description": "Creates a new entry through an `INSERT INTO` query.",
        "operationId": "addRestaurant",
        "requestBody": {
          "description": "Example Restaurant",
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "restaurant_name": {
                    "type": "string",
                    "description": "Name of the restaurant to add.",
                    "example": "Shawarma Palace"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful POST.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "string",
                      "description": "Confirmation of successful post",
                      "example": "Commit Successful!"
                    },
                    "status": {
                      "type": "integer",
                      "description": "status code returned by request.",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "type of error",
                      "example": "Bad Request!"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 400
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "restaurant"
        ],
        "summary": "Add proximity info to restaurant.",
        "description": "Add proximity info to restaurant through an `UPDATE` query which takes the proximity to and delivery time from a restaurant you've ordered from.",
        "requestBody": {
          "description": "Example Restaurant",
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "restaurant_name": {
                    "type": "string",
                    "description": "Name of the restaurant to add.",
                    "example": "Shawarma Palace"
                  },
                  "proximity": {
                    "type": "number",
                    "description": "Distance in km from the restaurant.",
                    "example": 2.5
                  },
                  "delivery_time": {
                    "type": "number",
                    "description": "Average time in minutes to get delivery from the restaurant.",
                    "example": 30
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful POST.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "string",
                      "description": "Confirmation of successful post",
                      "example": "Commit Successful!"
                    },
                    "status": {
                      "type": "integer",
                      "description": "status code returned by request.",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "type of error",
                      "example": "Bad Request!"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 400
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/dishes": {
      "get": {
        "tags": [
          "dishes"
        ],
        "summary": "Read all dishes, sorted by restaurant.",
        "description": "Runs a `SELECT` query against the tables for dishes, likes, and tags to pull all info related to each dish and groups them by restaurant. This call can optionally take query parameters to filter the data depending on who wants to eat or what dietary restrictions need to be followed (i.e. filtering by tag or user names).",
        "parameters": [
          {
            "name": "user",
            "in": "query",
            "description": "Name of a user who has to like the dishes being queried for.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Connor"
            }
          },
          {
            "name": "tag",
            "in": "query",
            "description": "Tag to be filtered for.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "Gluten-Free"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Dishes"
                },
                "example": {
                  "Shawarma Palace": {
                    "dishes": {
                      "Chicken Shawarma Wrap": {
                        "likes": [
                          "Connor"
                        ],
                        "tags": [
                          "Halal"
                        ]
                      },
                      "Beef Shawarma Wrap": {
                        "likes": null,
                        "tags": [
                          "Halal"
                        ]
                      }
                    }
                  },
                  "status": 200
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "dishes"
        ],
        "description": "Adds a new dish using an `INSERT INTO` command which takes the new dish name, and an existing restaurant, as input. Fails if there is an attempt to post a dish that already exists.",
        "requestBody": {
          "description": "Info to create a dish (i.e. Name and name of restaurant).",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "dish_name": {
                    "type": "string",
                    "description": "The name of the dish",
                    "example": "Chicken Shawarma Wrap"
                  },
                  "restaurant_name": {
                    "type": "string",
                    "description": "The name of the restaurant",
                    "example": "Shawarma Palace"
                  }
                }
              }
            }
          }
        },
        "summary": "Add a new dish",
        "responses": {
          "200": {
            "description": "Successful Post.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "string",
                      "description": "successful POST confirmation.",
                      "example": "Commit Successful!"
                    },
                    "status": {
                      "type": "integer",
                      "description": "Status code.",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "type of error",
                      "example": "Bad Request!"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 400
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/restaurants/{restaurant_name}/{dish_name}": {
      "delete": {
        "tags": [
          "dishes"
        ],
        "description": "Deletes a dish from a given restaurant using a `DELETE FROM` query on the `restaurant_name` and `dish_name` provided.",
        "summary": "Removes a dish.",
        "parameters": [
          {
            "name": "restaurant_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Shawarma Palace"
            }
          },
          {
            "name": "dish_name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Chicken Shawarma Wrap"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Delete.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "string",
                      "description": "successful DELETE confirmation.",
                      "example": "Deleted Chicken Shawarma Wrap successfully."
                    },
                    "status": {
                      "type": "integer",
                      "description": "Status code.",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "type of error",
                      "example": "Bad Request!"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 400
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/likes": {
      "get": {
        "tags": [
          "likes"
        ],
        "parameters": [
          {
            "name": "restaurant_name",
            "in": "query",
            "description": "Name of the restaurant to read likes from.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Shawarma Palace"
            }
          },
          {
            "name": "dish_name",
            "in": "query",
            "description": "Name of the dish.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "Chicken Shawarma Wrap"
            }
          }
        ],
        "summary": "Read likes for a given dish at a restaurant.",
        "description": "Runs a `SELECT *` query against the likes table using restaurant_name and dish_name as parameters.",
        "responses": {
          "200": {
            "description": "Successful Read.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Likes"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "type of error",
                      "example": "Bad Request!"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 400
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "likes"
        ],
        "description": "Runs a `INSERT INTO` query on the likes table using restaurant_name, dish_name and user_name as parameters. This is what runs when the user hits a like button.",
        "requestBody": {
          "description": "Info to specify the dish being liked.",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "dish_name": {
                    "type": "string",
                    "description": "The name of the dish",
                    "example": "Chicken Shawarma Wrap"
                  },
                  "restaurant_name": {
                    "type": "string",
                    "description": "The name of the restaurant",
                    "example": "Shawarma Palace"
                  },
                  "user_name": {
                    "type": "string",
                    "description": "The user who likes the dish",
                    "example": "Connor"
                  }
                }
              }
            }
          }
        },
        "summary": "Mechanism for a user to like a dish at a given restaurant.",
        "responses": {
          "200": {
            "description": "Successful Post.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "string",
                      "description": "successful POST confirmation.",
                      "example": "Connor liked Chicken Shawarma Wrap at Shawarma Palace."
                    },
                    "status": {
                      "type": "integer",
                      "description": "Status code.",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "type of error",
                      "example": "Bad Request!"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 400
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{user_name}": {
      "delete": {
        "tags": [
          "users"
        ],
        "parameters": [
          {
            "name": "user_name",
            "required": true,
            "in": "path",
            "description": "The name of the user in the application.",
            "schema": {
              "type": "string",
              "example": "Connor"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Delete.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "type": "string",
                      "description": "successful DELETE confirmation.",
                      "example": "Removed user Test User"
                    },
                    "status": {
                      "type": "integer",
                      "description": "Status code.",
                      "example": 200
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request.",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "error": {
                      "type": "string",
                      "description": "type of error",
                      "example": "Bad Request!"
                    },
                    "status": {
                      "type": "integer",
                      "format": "int64",
                      "example": 400
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Restaurant": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "example": 1
          },
          "restaurant_name": {
            "type": "string",
            "description": "name of the restaurant",
            "example": "Shawarma Palace"
          },
          "delivery_time": {
            "type": "number",
            "description": "How long it typically takes a delivery to arrive.",
            "example": 30.5
          },
          "proximity": {
            "type": "number",
            "description": "Distance in KM to restaurant.",
            "example": 2.5
          },
          "tags": {
            "type": "object",
            "description": "Other data to filter restaurants by.",
            "example": {
              "Type": "Shawarma"
            }
          }
        }
      },
      "Restaurants": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Restaurant"
        }
      },
...
      "Dishes": {
        "type": "object",
        "properties": {
          "restaurant": {
            "type": "object",
            "properties": {
              "dishes": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Dish"
                }
              }
            }
          }
        }
      },
...
      "Likes": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Like"
            }
          }
        }
      },
...
    }
  }
}